removed points entirely from the usecase diagram
points are a time unit. which the user could define by him self based on an hour. However that can completle change the leveling funktion. Additionally keyholders and lockees are used to dealing with time instead of points so I will standardise the points per hour to match one point per minute.